+++ src/HeaderCell.tsx
@@ -6,7 +6,7 @@ import { HeaderRowProps } from './HeaderRow';
 import SortableHeaderCell from './headerCells/SortableHeaderCell';
 import ResizableHeaderCell from './headerCells/ResizableHeaderCell';
 
-type SharedHeaderRowProps<R> = Pick<HeaderRowProps<R, never>,
+type SharedHeaderRowProps<R, SR> = Pick<HeaderRowProps<R, never, SR>,
   | 'sortColumn'
   | 'sortDirection'
   | 'onSort'
@@ -16,22 +16,22 @@ type SharedHeaderRowProps<R> = Pick<HeaderRowProps<R, never>,
   | 'draggableHeaderCell'
 >;
 
-export interface HeaderCellProps<R> extends SharedHeaderRowProps<R> {
-  column: CalculatedColumn<R>;
+export interface HeaderCellProps<R, SR> extends SharedHeaderRowProps<R, SR> {
+  column: CalculatedColumn<R, SR>;
   lastFrozenColumnIndex: number;
   scrollLeft: number | undefined;
-  onResize(column: CalculatedColumn<R>, width: number): void;
+  onResize(column: CalculatedColumn<R, SR>, width: number): void;
   onAllRowsSelectionChange(checked: boolean): void;
 }
 
-export default function HeaderCell<R>({
+export default function HeaderCell<R, SR>({
   height,
   column,
   allRowsSelected,
   onAllRowsSelectionChange,
   scrollLeft,
   ...props
-}: HeaderCellProps<R>) {
+}: HeaderCellProps<R, SR>) {
   function getCell() {
     if (!column.headerRenderer) return column.name;